refactor(lib/file): changed how ManagedObject works
ManagedObject was a CRTP originally designed to help the implementation of the
B+ tree. We reused it for storing general objects. However, it has several
issues:

- Every subclass of ManagedObject had to store its own file storage, and doing
  this requires writing file::File<sizeof(Foo)> foos {"foos"}; over and over,
  which is not great.
- Furthermore, when using subclasses of ManagedObject, the caller needs to pass
  the file object to the API, like Users::get(users, 1). It's a strange and
  pointless repetition.
- Efficiently storing models requires that models declare char _start[0] and
  char _end[0], which aside from being a hack, is ugly.

This commit rewrote the sequential file storage subsystem to address these
issues. For the purpose of not breaking B+ tree code, a copy of the original
ManagedObject is made under lib/file/internal/file.h. The new API extends the
user-provided base class rather than being a base class. The name ManagedObject
fits in the context Foo : public ManagedObject, but not Foo =
ManagedObject<FooBase>. Therefore, its name is changed to simply Managed.

The new API takes the approach of convention over configuration. You need to
specify the filename in the base class under the special static constexpr const
char *filename variable. Failure to provide this value would lead to a compile
error.

Other APIs are also modified to accompany this change; notably Index, which no
longer need to pass in the file object of the model, since the object needed is
already innate of the model class.