feat(node): introduce Node-API integration
This commit introduces Node-API integration to the project, among other
things. It is generally not recommended to make such a lot of changes in
a single commit, but these changes are somewhat inseparable, due to
previous oversight and lack of testing.

The main part is the src/node.cpp entrypoint. In contrary of making a
pipe (unix domain socket, tcp socket etc), we make the main C++ logic
available to Node.js by wrapping the project as an native node module
(aka native addon). This is achieved by using the new Node-API and the
cmake-js build tool.

To build the main node module:

1. Install Node.js (my version is 17.6.0), CMake and Make
2. Install cmake-js: npm i cmake-js -g
3. Install dependencies: npm i, or yarn
4. Run the build script: bin/build-node

The module built would be located at build/Release/ticket.node.

Currently the module exports a native function receiving an array of
strings as the arguments. This design is somewhat bad and is subject
to future changes.

In order to make this happen, the following changes are also made:

- datetime.cpp is implemented. This is neccessary for the project to
  compile.
- user.cpp is implemented as an example and a small manual test drive.
- command::dispatch now return Results instead of void. This allows the
  seperation of logic and presentation, so that we could write two
  presentations, one on standard output (aka std::cout), and the other
  on Node. The presentation layer is in src/response.{h,cpp}, and is not
  final; changes are expected to be made.
- Several small fixes and minor features in the library.