feat(node): proper packaging of C++ APIs
Previously, the C++ API is only available by passing the arguments
as an array of strings. This is ugly and insecure. Dev experience
sucks. This commit introduces a new API by wrapping each command
as a JS function, and uses JS objects to pass the arguments. A
TypeScript declaration is included, though it is suboptimal as all
commands return the Response sum type.

Example:

- Old API: ticket([ 'add_user', '-u', 'yyu' ])
- New API: ticket.addUser({ username: 'yyu' })