feat(lib): write cache for the lru cache
Previously, the LRU cache did not cache write operations. Data is
written to the file when the set function is executed. This caused
a lot of unnecessary writes, especially in hot spots. This commit
changes the LRU cache so that the data is only written to the file
when the entry is cleared from the cache.