refactor(parser): split out the run declarations
Previously, the run() overloads of the commands are declared in
parser.h. This caused a dependency loop in development:

- parser.h -> response.h (for Response)
- response.h -> order.h (for Order etc.)
- order.h -> train.h (for Train etc.)
- train.h -> parser.h (for commands etc.)

This commit breaks the first dependency by moving run(), the part
that requires Response, out to a new file, run.h. This header is
included only by .cpp files, thus avoiding further dependency loops.